[Opinel] Moved cli_parser to core
