+++ ScoutSuite/core/cli_parser.py
