+++ ScoutSuite/core/console.py
