+++ ScoutSuite/core/exceptions.py
@@ -3,7 +3,7 @@
 Exceptions handling
 """
 
-from core.console import printDebug
+from ScoutSuite.core.console import printDebug
 
 from ScoutSuite import EXCEPTIONS
 from ScoutSuite.output.js import JavaScriptReaderWriter
