+++ ScoutSuite/core/processingengine.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from core.console import printDebug, printError, printException
+from ScoutSuite.core.console import printDebug, printError, printException
 from opinel.utils.globals import manage_dictionary
 
 from ScoutSuite.core.utils import recurse
