+++ ScoutSuite/core/rule.py
@@ -4,7 +4,7 @@ import json
 import re
 
 from opinel.utils.fs import read_ip_ranges
-from core.console import printError
+from ScoutSuite.core.console import printError
 
 from ScoutSuite.utils import format_service_name
