+++ ScoutSuite/providers/aws/configs/regions.py
@@ -13,7 +13,7 @@ except ImportError:
     from queue import Queue
 
 from opinel.utils.aws import build_region_list, connect_service, get_aws_account_id, get_name, handle_truncated_response
-from core.console import printException, printInfo
+from ScoutSuite.core.console import printException, printInfo
 from opinel.utils.globals import manage_dictionary
 
 from ScoutSuite.providers.base.configs import resource_id_map
