+++ ScoutSuite/providers/aws/provider.py
@@ -6,7 +6,7 @@ import sys
 
 try:
     from opinel.utils.aws import get_aws_account_id, get_partition_name
-    from core.console import configPrintException, printInfo, printDebug
+    from ScoutSuite.core.console import configPrintException, printInfo, printDebug
     from opinel.utils.credentials import read_creds
     from opinel.utils.globals import check_requirements
     from opinel.utils.profiles import AWSProfiles
@@ -16,7 +16,7 @@ except Exception as e:
     print(e)
     sys.exit(42)
 
-from core.console import printDebug, printError, printException, printInfo
+from ScoutSuite.core.console import printDebug, printError, printException, printInfo
 from opinel.utils.globals import manage_dictionary
 
 from ScoutSuite.providers.aws.configs.services import AWSServicesConfig
