+++ ScoutSuite/providers/base/provider.py
@@ -7,7 +7,7 @@ import json
 import sys
 import copy
 
-from core.console import printException, printInfo
+from ScoutSuite.core.console import printException, printInfo
 
 from ScoutSuite import __version__ as scout2_version
 from ScoutSuite.providers.base.configs.browser import get_object_at
