+++ ScoutSuite/providers/gcp/services/cloudresourcemanager.py
@@ -2,7 +2,7 @@
 
 from ScoutSuite.providers.gcp.configs.base import GCPBaseConfig
 
-from core.console import printException
+from ScoutSuite.core.console import printException
 
 
 class CloudResourceManager(GCPBaseConfig):
